fix some clippy errors
I'm not sure how these slipped through before, it's possibly caused by
upgrading rules_rust and/or rust?

+++ engine/behavior.rs
@@ -257,7 +257,7 @@ impl TriggerType for AgentPlanCommuteToWork {
             // we had this originally, but it's tougher with parallelism
             engine.trigger_queue.push(
                 AgentPlanCommuteHome { agent: id },
-                start_time as u64 + Time::new::<hour>(8).value,
+                start_time + Time::new::<hour>(8).value,
             );
         }
